Rename odc-index package to odc-stac #330
