+++ apps/dc_tools/setup.cfg
@@ -29,7 +29,7 @@ install_requires =
     toolz
     pyyaml
     datacube
-    odc_index
+    odc_stac
     odc_io
     odc_aws
     odc_aio
