+++ libs/algo/setup.cfg
@@ -26,7 +26,7 @@ install_requires =
     xarray
     numpy
     toolz
-    odc-index
+    odc-stac
     datacube
     scikit-image
     dask_image
