+++ libs/dscache/setup.cfg
@@ -20,7 +20,7 @@ python_requires = >=3.6
 tests_require = pytest
 install_requires =
     datacube
-    odc-index
+    odc-stac
     zstandard
     lmdb
     click
