+++ libs/index/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = odc-index
-description = Datacube index prototypes/tools
-version = attr: odc.index._version.__version__
+name = odc-stac
+description = Tooling for converting STAC metadata to ODC data model
+version = attr: odc.stac._version.__version__
 author = Open Data Cube
 author_email =
 maintainer = Open Data Cube
