+++ libs/stats/setup.cfg
@@ -20,7 +20,7 @@ python_requires = >=3.6
 tests_require = pytest
 install_requires =
     datacube
-    odc_index
+    odc_stac
     odc_dscache
     odc_algo
     odc_aws
