+++ libs/ui/setup.cfg
@@ -21,7 +21,7 @@ tests_require = pytest
 install_requires =
     ipywidgets
     ipyleaflet
-    odc_index
+    odc_stac
     odc_algo
     jupyter_ui_poll
